# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tinder Lite API",
        "description": "API documentation for Tinder-like app",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/people/{id}/like": {
            "post": {
                "tags": [
                    "People"
                ],
                "summary": "Like a person",
                "operationId": "bc2adad842af56d5dadcd5a1059059d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person liked"
                    }
                }
            }
        },
...
    },
...
}